Wayland: Add initial support for drawing tablets
This adds very basic support for tablet motion and tools to GDK using the
Wayland backend. Tablet motion is multiplexed through the same cursor the
pointer uses so we don't have to break everything by having more then one
cursor. The cursor image is also multiplexed, and GDK switches the image between
whichever mouse/tablet cursor moved last, and hides the images on all of the
others in order to give the impression that there is only ever one active
pointer on a GTK+ application.

                                    Changes
- When adding a tablet device, add it to the universal list of devices
- Remove some unused variables
- Set wl_tablet in GdkWaylandDeviceTabletPair when initializing tablet
- Always update the cursor device when a multiplexed cursor device leaves the
  proximity of a surface to prevent potential segfaults
- Handle axis frames and send motion events when we receive them, instead of
  sending the motion events right when we get the axis update from the
  compositor. This is needed if we want to prevent redundant motion events once
  we add support for pressure, distance and tilt.
- tablet_handle_removed(): Don't use wl_tablet_get_user_data() since the 'data'
  pointer provides an identical value
- Fix typo in one of the comments I added
- Remove unneeded comment in tablet_listener
- Separate code for making cursor device image blank into it's own function
- Use g_list_append() as opposed to g_list_prepend(), since we don't really care
  about the ordering in the lists