Wayland: Ignore other cursor devices when button is held down
When we're multiplexing cursor devices, there's an (unlikely) chance that the
user could accidentally move the pointer with a tablet device while they're
dragging something with the cursor. To prevent this, we lock the cursor device
to the mouse while any mouse buttons are held down, and release the lock when
the buttons are released. During the lock, all tablet pointers are drawn as
blank, and all of the events from the compositor for the tablet devices are not
passed onto the GDK client.