Wayland: Add GdkWaylandDeviceTool
Because the axes that can be provided by a tool in wayland are per-tool, we have
to be able to store the axis capabilities of each tool along with the tool
itself. GdkWaylandDeviceTool does this by inheriting GdkDeviceTool and adding
another field for storing the axis capabilities reported by the compositor.