Wayland: Report axis capabilities of tablet tools
Axis capabilities in Wayland are per-tool. This means that whenever a new tool
comes into proximity, the available capabilities of both the tablet and
inherently the master pointer change. As such, we update the axis capabilities
of the tablet to match those of the tool, and have the master pointer device
"mimic" the axis capabilities of whichever pointer device the on-screen pointer
corresponds to.

                                    Changes
- Keep track of the index values of each axis, so we don't have to look them up
  every time we try to update one of their values.