wayland: Always send cursor event/leave events
Originally we would only send cursor leave/enter events when entering a window,
since we were multiplexing the cursors. Since tablets have their own cursors
now, we need to always send enter/leave events. In addition, since this gets rid
of the need to keep track of how many tablets are on a single surface, the code
for incrementing/decrementing the tablet focus count on a surface has been
removed.